requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data processing
opencv-python>=4.8.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
pyyaml>=6.0
seaborn>=0.12.0

# MLOps
mlflow>=2.9.0

# Serving (optional, for existing pipeline compatibility)
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6
